+++ docs/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog for Typopo
 
-## 2016-03-14
+## 2016-03-13
  * 0.0.10
- * remove accidental uPerRCase
+ * correct accidental uPerRCase
+ * make typopo as an NPM package (thanks to Matus Duchon)
 
 ## 2016-02-20
  * 0.0.9